feat(layout): split PartnerGyms into Marketing + Contact sections; reorder, drop WhatThalosIs + FounderNote
